+++ packages/data-source/src/lib/types/events.ts
@@ -1,0 +1,6 @@
+import { DataSourceController } from "../classes";
+
+/**
+ * On data changed event callback function
+ */
+export type OnDataChangedCallback<T> = (data: T[], controller: DataSourceController<T>) => void;
