+++ packages/data-source/src/lib/types/response.ts
@@ -1,0 +1,8 @@
+/**
+ * Function for performing a fetch call
+ */
+export type FetchResponseAsync = (signal?: AbortSignal) => Promise<Response>;
+/**
+ * Function for parsing the response
+ */
+export type ResponseParserAsync<T> = (Response: Response) => Promise<T[]>;
